requirements.txt
prometheus_client==0.20.0
speedtest-cli==2.1.3
